refactor : 문자열 계산기 패키지 정리
