+++ src/test/java/calculator/operator/OperatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.operator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
