+++ src/test/java/calculator/service/StringCalculatorTest.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.service;
 
 import static org.assertj.core.api.Assertions.assertThat;
